+++ src/content/orders.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useUser } from "@propelauth/nextjs/client";
 import { Button, Spinner, DatePicker, Card, CardBody } from "@nextui-org/react";
-import { useState, useEffect, useCallback } from "react";
+import { useState, useEffect, useCallback, useMemo } from "react";
 import React from "react";
 import { saveAs } from "file-saver";
 import { demoFlag, not_products } from "@/helpers/utils";
@@ -31,6 +31,12 @@ import {
   getOrders,
 } from "@/helpers/request";
 import { useQuery, useQueryClient } from "react-query";
+import {
+  filterOrdersByStatus,
+  filterOrdersByCategory,
+  filterBySearch,
+  filterOrdersByComponent,
+} from "@/helpers/filters";
 
 export default function OrdersPage() {
   const { user } = useUser();
@@ -40,7 +46,6 @@ export default function OrdersPage() {
   const [startDate, setStartDate] = useState(
     now(getLocalTimeZone()).subtract({ weeks: 1 })
   ); // Default to a week ago
-  const [filteredOrders, setFilteredOrders] = useState<any[]>([]);
   const [selectedMenuKeys, setSelectedMenuKeys] = useState<any>(
     new Set(["All"])
   );
@@ -185,6 +190,25 @@ export default function OrdersPage() {
       </Button>
     );
   };
+
+  const filteredOrders = useMemo(() => {
+    if (!orders) return [];
+    let filtered = filterOrdersByStatus(orders, selectedStatusKeys);
+    filtered = filterOrdersByCategory(filtered, selectedMenuKeys, products);
+    filtered = filterOrdersByDate(filtered, deliveryDate);
+    filtered = filterOrdersByComponent(filtered, selectedComponent);
+    filtered = filterBySearch(filtered, searchTerm);
+    return filtered;
+  }, [
+    orders,
+    products,
+    selectedStatusKeys,
+    selectedMenuKeys,
+    deliveryDate,
+    selectedComponent,
+    searchTerm,
+  ]);
+
   useEffect(() => {
     if (orders.length > 0) {
       const hasComposite = checkForCompositeProducts(orders);
@@ -196,156 +220,15 @@ export default function OrdersPage() {
     }
   }, [orders]);
 
-  const filterOrdersByComponent = (
-    orders: any[],
-    selectedKeys: Set<string>
-  ) => {
-    return orders
-      .map((order) => {
-        const includedProducts = new Set();
-        return {
-          ...order,
-          line_items: order.line_items.filter((item: any) => {
-            if (selectedKeys.has("All")) {
-              return true;
-            }
-            if (item.composite_parent) {
-              const compositeData = item.meta_data.find(
-                (meta: any) => meta.key === "_composite_data"
-              );
-              if (compositeData && compositeData.value) {
-                const matchedComponent = Object.values(
-                  compositeData.value
-                ).find(
-                  (component: any) =>
-                    component.product_id === item.product_id &&
-                    selectedKeys.has(component.title)
-                );
-                if (
-                  matchedComponent &&
-                  !includedProducts.has(item.product_id)
-                ) {
-                  includedProducts.add(item.product_id);
-                  return true;
-                }
-              }
-            }
-            return false;
-          }),
-        };
-      })
-      .filter(
-        (order) => selectedKeys.has("All") || order.line_items.length > 0
-      );
-  };
-
   const handleSearch = useCallback((term: string) => {
     setSearchTerm(term);
   }, []);
-
-  useEffect(() => {
-    if (orders) {
-      const filtered = getFilteredOrders(
-        orders,
-        selectedMenuKeys,
-        selectedStatusKeys,
-        selectedComponent
-      );
-      setFilteredOrders(filtered);
-    }
-  }, [
-    selectedMenuKeys,
-    selectedStatusKeys,
-    orders,
-    deliveryDate,
-    searchTerm,
-    selectedComponent,
-  ]);
-
-  const nameSearch = (orders: any[], searchTerm: string) => {
-    // Extract the name search terms (after "name:")
-    const nameSearchTerms = searchTerm
-      .slice(5)
-      .split(",")
-      .map((term: string) => term.trim().toLowerCase());
-
-    // Filter orders based on the full name derived from first_name and last_name
-    const filteredOrders = orders.filter((order) => {
-      const fullName =
-        `${order.billing.first_name} ${order.billing.last_name}`.toLowerCase();
-      // Check if any of the search terms match the full name
-      return nameSearchTerms.some((term) => fullName.includes(term));
-    });
-
-    return filteredOrders;
-  };
 
   const handleDownloadDeliveryList = () => {
     // hardcoding
     const areaZipcodeMap = org.zipcodeMap;
     console.log("Area Zipcode Map: ", areaZipcodeMap);
     deliveryList(filteredOrders, areaZipcodeMap);
-  };
-
-  const getFilteredOrders = (
-    orders: any[],
-    selectedMenuKeys: Set<string>,
-    selectedStatusKeys: Set<string>,
-    selectedComponent: Set<string>
-  ) => {
-    let filtered = filterOrdersByStatus(orders, selectedStatusKeys);
-    filtered = filterOrdersByCategory(filtered, selectedMenuKeys);
-    filtered = filterOrdersByDate(filtered, deliveryDate);
-    filtered = filterOrdersByComponent(filtered, selectedComponent);
-    filtered = filterBySearch(filtered, searchTerm);
-    return filtered;
-  };
-
-  const filterBySearch = (orders: any[], searchTerm: string) => {
-    if (searchTerm.toLowerCase().startsWith("name:")) {
-      return nameSearch(orders, searchTerm);
-    } else {
-      // Regular filter based on id field (unchanged)
-      const filteredOrders = orders.filter((order) =>
-        order.id.toString().toLowerCase().includes(searchTerm.toLowerCase())
-      );
-      return filteredOrders;
-    }
-  };
-
-  const filterOrdersByCategory = (orders: any[], selectedKeys: Set<string>) => {
-    return orders
-      .map((order) => ({
-        ...order,
-        line_items: order.line_items.filter((item: any) => {
-          if (selectedKeys.has("All")) {
-            return true;
-          }
-          const associatedProduct = products.find(
-            (product: any) => product.id === item.product_id
-          );
-          const itemCategories =
-            associatedProduct?.categories?.map(
-              (category: any) => category.name
-            ) || [];
-          return Array.from(selectedKeys).every((selectedCategory: any) =>
-            itemCategories.includes(selectedCategory)
-          );
-        }),
-      }))
-      .filter(
-        (order) => selectedKeys.has("All") || order.line_items.length > 0
-      );
-  };
-
-  const filterOrdersByStatus = (orders: any[], selectedKeys: Set<string>) => {
-    // Check if "All" is in the selectedKeys
-    if (selectedKeys.has("All")) {
-      return orders;
-    }
-
-    // Filter the orders where the status is one of the selected keys
-    return orders.filter((order) => selectedKeys.has(order.status));
   };
 
   const renderComponentFilterDropdown = () => {
